Moved edit button in catalog to actions menu
